requirements.txt
cmake
easydict~=1.9
numpy
numba
scipy
scikit-image
scikit-learn
tqdm
PyYAML
open3d
cython
tensorboardX
einops
timm
icecream
h5py
pyquaternion
efficientnet_pytorch==0.7.0
torch-scatter
torch-cluster
ftfy
regex